requirements.txt
numpy<2.0
pandas<2.1
gymnasium
ray[rllib]
torch
requests
fastapi
uvicorn
supabase
python-dotenv
google-genai
python-multipart
